NetRocks aestetical improvements (fix #618)
