+++ far2l/plugin.hpp
@@ -1946,6 +1946,59 @@ typedef int (WINAPI *FAREXECUTE_LIBRARY)(const wchar_t *Library, const wchar_t *
 typedef void (WINAPI *FARDISPLAYNOTIFICATION)(const wchar_t *action, const wchar_t *object);
 typedef int (WINAPI *FARDISPATCHNTRTHRDCALLS)();
 
+enum BOX_DEF_SYMBOLS
+{
+	BS_X_B0,          // 0xB0
+	BS_X_B1,          // 0xB1
+	BS_X_B2,          // 0xB2
+	BS_V1,            // 0xB3
+	BS_R_H1V1,        // 0xB4
+	BS_R_H2V1,        // 0xB5
+	BS_R_H1V2,        // 0xB6
+	BS_RT_H1V2,       // 0xB7
+	BS_RT_H2V1,       // 0xB8
+	BS_R_H2V2,        // 0xB9
+	BS_V2,            // 0xBA
+	BS_RT_H2V2,       // 0xBB
+	BS_RB_H2V2,       // 0xBC
+	BS_RB_H1V2,       // 0xBD
+	BS_RB_H2V1,       // 0xBE
+	BS_RT_H1V1,       // 0xBF
+	BS_LB_H1V1,       // 0xC0
+	BS_B_H1V1,        // 0xC1
+	BS_T_H1V1,        // 0xC2
+	BS_L_H1V1,        // 0xC3
+	BS_H1,            // 0xC4
+	BS_C_H1V1,        // 0xC5
+	BS_L_H2V1,        // 0xC6
+	BS_L_H1V2,        // 0xC7
+	BS_LB_H2V2,       // 0xC8
+	BS_LT_H2V2,       // 0xC9
+	BS_B_H2V2,        // 0xCA
+	BS_T_H2V2,        // 0xCB
+	BS_L_H2V2,        // 0xCC
+	BS_H2,            // 0xCD
+	BS_C_H2V2,        // 0xCE
+	BS_B_H2V1,        // 0xCF
+	BS_B_H1V2,        // 0xD0
+	BS_T_H2V1,        // 0xD1
+	BS_T_H1V2,        // 0xD2
+	BS_LB_H1V2,       // 0xD3
+	BS_LB_H2V1,       // 0xD4
+	BS_LT_H2V1,       // 0xD5
+	BS_LT_H1V2,       // 0xD6
+	BS_C_H1V2,        // 0xD7
+	BS_C_H2V1,        // 0xD8
+	BS_RB_H1V1,       // 0xD9
+	BS_LT_H1V1,       // 0xDA
+	BS_X_DB,          // 0xDB
+	BS_X_DC,          // 0xDC
+	BS_X_DD,          // 0xDD
+	BS_X_DE,          // 0xDE
+	BS_X_DF,          // 0xDF
+};
+
+
 typedef struct FarStandardFunctions
 {
 	int StructSize;
@@ -1964,7 +2017,8 @@ typedef struct FarStandardFunctions
 	FARSTDSNPRINTF             snprintf;
 	// </C&C++>
 
-	DWORD_PTR                  Reserved[8];
+	DWORD_PTR                  Reserved[7];
+	const WCHAR *              BoxSymbols; // indexed via BOX_DEF_SYMBOLS
 
 	FARSTDLOCALISLOWER         LIsLower;
 	FARSTDLOCALISUPPER         LIsUpper;
